tests/apps/Healthcare-voice-agent/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
langgraph
langchain
openai
google-generativeai
pydantic
python-dotenv
requests
tqdm
numpy
langchain-openai
duckduckgo-search
langchain-community

